requirements.txt
discord.py 
requests 
